feat: implementa label, suporte ao enter e validação de campo vazio, e renomeia o componente para DynamicListInput
