+++ src/components/index.ts
@@ -30,6 +30,7 @@ import ColorPicker from './ColorPicker.vue';
 import DashboardCard from './DashboardCard.vue';
 import DateInput from './DateInput.vue';
 import DialogModal from './DialogModal.vue';
+import DynamicInputList from './DynamicInputList.vue';
 import Divider from './Divider.vue';
 import Dropdown from './Dropdown.vue';
 import DropdownButton from './DropdownButton.vue';
@@ -60,7 +61,6 @@ import MobileNavbar from './MobileNavbar.vue';
 import MobileNavigation from './MobileNavigation.vue';
 import Modal from './Modal.vue';
 import MultiFileInput from './MultiFileInput.vue';
-import MultiInput from './MultiInput.vue';
 import Multiselect from './Multiselect.vue';
 import NavBar from './NavBar.vue';
 import NumberInput from './NumberInput.vue';
@@ -184,7 +184,7 @@ export default {
 		app.component('CdsMobileNavigation', MobileNavigation);
 		app.component('CdsModal', Modal);
 		app.component('CdsMultiselect', Multiselect);
-		app.component('CdsMultiInput', MultiInput);
+		app.component('CdsDynamicInputList', DynamicInputList);
 		app.component('CdsMultiFileInput', MultiFileInput);
 		app.component('CdsNavBar', NavBar);
 		app.component('CdsNumberInput', NumberInput);
